Initial commit
So far:
Links are rods with length and possibly an angle with a connection
Bases are conceptually just another Link with a position
Mechanisms are Links and a Connection which defines the Loop

The Loop is a System of Equations which should result in the 0-Vector.
For this the mathjs Library was added as dependency to derivate and to
create the Jacobi-Matrix, which is the next step.