Switch to WebPack and calculate q (probably way off)
Tried to implement google-closure-compiler, but that did not
import other modules correctly.
Tried webpack and it works. mathjs seems to do it's job.

Now it is time to reduce all the visual helpers and try to get
q to 0. First check the results, because q seems to be way off.